+++ addons/yat/src/interfaces/ICommand.cs
@@ -15,7 +15,34 @@ public partial interface ICommand
 	/// </summary>
 	/// <param name="data">The data passed to the command.</param>
 	/// <returns>The result of the command execution.</returns>
-	public ECommandResult Execute(CommandData data);
+	public CommandResult Execute(CommandData data);
+
+	public virtual CommandResult Success(string message = null) =>
+		new(ECommandResult.Success, message);
+
+	public virtual CommandResult Failure(string message = null) =>
+		new(ECommandResult.Failure, message);
+
+	public virtual CommandResult InvalidArguments(string message = null) =>
+		new(ECommandResult.InvalidArguments, message);
+
+	public virtual CommandResult InvalidCommand(string message = null) =>
+		new(ECommandResult.InvalidCommand, message);
+
+	public virtual CommandResult InvalidPermissions(string message = null) =>
+		new(ECommandResult.InvalidPermissions, message);
+
+	public virtual CommandResult InvalidState(string message = null) =>
+		new(ECommandResult.InvalidState, message);
+
+	public virtual CommandResult NotImplemented(string message = null) =>
+		new(ECommandResult.NotImplemented, message);
+
+	public virtual CommandResult UnknownCommand(string message = null) =>
+		new(ECommandResult.UnknownCommand, message);
+
+	public virtual CommandResult UnknownError(string message = null) =>
+		new(ECommandResult.UnknownError, message);
 
 	public virtual string GenerateCommandManual()
 	{