+++ CHANGELOG.md
@@ -9,10 +9,12 @@ All notable changes to this project will be documented in this file.
 -   Options -id & -name to ToggleAudio command.
 -   More results to ECommandResult.
 -   CommandResult record.
+-   Virtual methods to ICommand that corresponds to the ECommandResult results.
 
 ### Changed
 
 -   Renamed CommandResult to ECommandResult.
+-   Commands now return CommandResult record.
 
 ## [1.23.0-beta 2024-02-13]